+++ templates/integrated/Source/Features/Members/SignUp/Mapper.cs
@@ -1,5 +1,4 @@
-﻿using MyProject;
-using Dom;
+﻿using Dom;
 
 namespace Members.Signup;