+++ templates/integrated/Source/Features/Members/SignUp/Request.cs
@@ -1,5 +1,4 @@
-﻿using MyProject;
-using FluentValidation;
+﻿using FluentValidation;
 
 namespace Members.Signup;