+++ templates/integrated/Source/Tests/Meta.cs
@@ -1,4 +1,3 @@
 ﻿global using FastEndpoints.Testing;
 global using FluentAssertions;
 global using Xunit;
-global using Xunit.Priority;