+++ templates/project/Tests/Meta.cs
@@ -4,4 +4,3 @@ global using FluentAssertions;
 global using Microsoft.Extensions.DependencyInjection;
 global using System.Net;
 global using Xunit;
-global using Xunit.Priority;